+++ Runtime/Common/Scripts/StereoConvergenceDetector.cs
@@ -22,8 +22,12 @@ using MagicLeap.MRTK.Settings;
 
 #if UNITY_OPENXR_1_9_0_OR_NEWER && MAGICLEAP_UNITY_SDK_2_0_0_OR_NEWER
 using UnityEngine.XR.OpenXR;
-using UnityEngine.XR.OpenXR.Features.MagicLeapSupport;
 using UnityEngine.XR.OpenXR.Features.Interactions;
+#if MAGICLEAP_UNITY_SDK_2_3_0_OR_NEWER
+using MagicLeap.OpenXR.Features;
+#else
+using UnityEngine.XR.OpenXR.Features.MagicLeapSupport;
+#endif
 #endif
 
 #if MAGICLEAP_UNITY_SDK_2_1_0_OR_NEWER